Update and fix several issues. Pin to pre-SDU. (#109)
* Update SDK and fix branchInfo lookup

* the branch info lookup isn't in the dialog anymore so handle it with some code
* update the SDK

* Use LT instead of size 0

Zero isn't supported anymore as a light sized environment
(with the updated SDK and version). Use LT.

* Use LT instead of size 0

Zero isn't supported anymore as a light sized environment
(with the updated SDK and version). Use LT.

* Pin version prior to SDU. Handle empty input better.

Starting with version 2019-04-30 we'll need to do more with SDU. The pinned version
still works as it used to.

* Pinned Disco prior to SDU
* Avoid Tone Analyzer w/ empty input
* Avoid NLU lang-detect errors by specifying English
* Handle branchInfo location lookup but forward to disco if no location was found.

* Remove sudo from .travis.yml

Fixes #102

* Update sample output image

Bad example doesn't work anymore.
Fix it in the image.

Fixes #108

* Update the Assistant json with one new example for demo

Added the "can use an atm in any city" example to the #Access
intent examples because it was used in the video and used to work.
Even though it is a little cryptic, the easiest fix is to add it
as an example and presto -- fixed.